introducing abstract base for ThumbnailGenerator
